Finish writing logic for performing basic calculations
Calculator functions are working as intended whenever the user clicks '=' or any other operator (assuming that there are any number inputs from the user beforehand). Further polishing
is needed for the logic as well as the overall design of the calculator. The project will later be revisited for said polishing but for now, the project is good as it is.